feat!: Two modifier keys can be used for hotkeys
BREAKING

- Existing user settings will fail to load and be cleared upon the first load.
  - Settings will be saved as usual afterwards.
  - Renamed some fields and added a lock type enum.

 Added

- Adds support for up to two modifier keys when registering the hotkey.
- Adds a file logger to troubleshoot issues.

Changed

- Split into three separate projects.
  - App for the main WPF application.
  - Native for Win32 API calls.
  - Sdk for models and main logic.

+++ src/AutoCursorLock.App/Extensions/KeyExtensions.cs
@@ -15,8 +15,8 @@ internal static class KeyExtensions
     /// </summary>
     /// <param name="key">The key to convert.</param>
     /// <returns>The Win32 virtual key code.</returns>
-    public static uint ToVirtualKey(this Key key)
+    public static int ToVirtualKey(this Key key)
     {
-        return (uint)KeyInterop.VirtualKeyFromKey(key);
+        return KeyInterop.VirtualKeyFromKey(key);
     }
 }
